database_grpc/db_grpc/pyproject.toml
[tool.poetry]
name = "db-grpc"
version = "0.1.0"
description = ""
authors = ["yash3004 <yashsehgal9253@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pymongo = "^4.8.0"
grpcio = "^1.65.1"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
